refactor(core): move Annotation to separate file
